requirements.txt
altgraph==0.17.3
certifi==2022.12.7
charset-normalizer==2.1.1
filelock==3.9.0
idna==3.4
imageio==2.28.0
Jinja2==3.1.2
MarkupSafe==2.1.2
mpmath==1.2.1
networkx==3.0
nibabel==5.1.0
numpy==1.24.3
opencv-python==4.7.0.72
packaging==23.1
pefile==2023.2.7
Pillow==9.3.0
pydicom==2.3.1
pyinstaller==5.10.1
pyinstaller-hooks-contrib==2023.2
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.1
pywin32-ctypes==0.2.0
requests==2.28.1
sympy==1.11.1
torch==2.0.0+cu118
torchaudio==2.0.1+cu118
torchvision==0.15.1+cu118
typing_extensions==4.4.0
urllib3==1.26.13
pytesseract~=0.3.10
matplotlib~=3.7.1